Added some changes

+++ org-mgmt-system/department-microservice/src/main/java/com/cloudcomp/ccoms/deptsvc/client/EmployeeClient.java
@@ -1,7 +1,7 @@
 package com.cloudcomp.ccoms.deptsvc.client;
 
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 
@@ -15,7 +15,7 @@ import com.cloudcomp.ccoms.deptsvc.model.Employee;
 public interface EmployeeClient {
     
     @GetMapping("/employee/dept/{deptId}")
-    public List<Employee> findEmpsByDeptId(@PathVariable("deptId") BigInteger deptId);
+    public List<Employee> findEmpsByDeptId(@PathVariable("deptId") Long deptId);
     
 }
